+++ CHANGELOG.md
@@ -9,6 +9,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [1.14.0] - 2026-02-26
+
+### Added
+
+- Set Neovim Leader to space
+- Add Neogit to Neovim with `<leader>gg` mapping
+- Add Neogit optional integrations via `telescope.nvim` (picker) and `codediff.nvim` (external diff backend)
+
 ## [1.13.1] - 2026-02-26
 
 ### Added
@@ -133,7 +141,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Scripts to manage the setup : install.sh and run-ansible.sh
 - Documentation in /docs with tasks to improve the setup, tips, etc.
 
-[Unreleased]: https://github.com/magoyette/dev-setup/compare/v1.13.1...HEAD
+[Unreleased]: https://github.com/magoyette/dev-setup/compare/v1.14.0...HEAD
+[1.14.0]: https://github.com/magoyette/dev-setup/compare/v1.13.1...v1.14.0
 [1.13.1]: https://github.com/magoyette/dev-setup/compare/v1.13.0...v1.13.1
 [1.13.0]: https://github.com/magoyette/dev-setup/compare/v1.12.1...v1.13.0
 [1.12.1]: https://github.com/magoyette/dev-setup/compare/v1.12.0...v1.12.1